pyproject.toml
[project]
name = "map-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "scipy>=1.17.1",
]
